oVirt API version is checked
If the engine's version is lower than required (recently 4.0),
an error message is shown, the oVirt polling is skipped and
all oVirt actions are redirected back to just Libvirt.

Such redirected actions will fail on an oVirt host now due to not-yet
implemented authentication to Libvirt which is required on an oVirt host.